backend/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
sqlalchemy==1.4.23
pymysql==1.0.2
mysqlclient==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
pydantic==1.10.2
python-dotenv==1.0.0
PyJWT==2.3.0
openpyxl==3.0.9
APScheduler==3.11.0
bcrypt==4.0.1
cryptography==44.0.2
pillow==11.1.0
reportlab==4.3.1
starlette==0.22.0
holidays==0.68
alembic==1.13.1